add weekly ticket logic

+++ src/api/controllers/user.js
@@ -1,4 +1,6 @@
 const { generateSign } = require('../../config/jwt');
+const { getWeekNum } = require('../../utils/weekCalc');
+const Ticket = require('../models/ticket');
 const User = require('../models/user');
 const bcrypt = require('bcrypt');
 
@@ -33,8 +35,29 @@ const login = async (req, res, next) => {
     const validPassword = bcrypt.compareSync(req.body.password, user?.password);
 
     if (user && validPassword) {
-      const token = generateSign(user._id);
-      return res.status(200).json({ message: 'Login exitoso', user, token });
+      const { prevLoginWeek, _id } = user;
+      const token = generateSign(_id);
+      const thisWeek = getWeekNum();
+      const firstLogin = !prevLoginWeek;
+      // si el último login fue hace una semana o es la primera vez que el usuario se loguea
+      if (thisWeek - prevLoginWeek === 1 || firstLogin) {
+        // genero un nuevo ticket y se lo asigno al usuario
+        const newTicket = new Ticket({ owner: _id });
+        const savedTicket = await newTicket.save();
+        user.tickets.push(savedTicket._id);
+      }
+      // reseteo loginWeek y actualizo los tickets
+      const updatedUser = await User.findByIdAndUpdate(
+        _id,
+        {
+          prevLoginWeek: thisWeek,
+          tickets: [...user.tickets],
+        },
+        { new: true }
+      );
+      return res
+        .status(200)
+        .json({ message: 'Login exitoso', updatedUser, token });
     } else {
       return res.status(401).json('Email o contraseña incorrectos.');
     }
@@ -53,11 +76,16 @@ const editUser = async (req, res, next) => {
 const deleteUser = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const deletedUser = await User.findByIdAndDelete(id);
-
-    return res
-      .status(200)
-      .json(`Usuario ${deletedUser.email} eliminado exitosamente`);
+    if (id === req.user.id || req.user.role === 'admin') {
+      const deletedUser = await User.findByIdAndDelete(id);
+      return res
+        .status(200)
+        .json(`Usuario ${deletedUser.email} eliminado exitosamente`);
+    } else {
+      return res
+        .status(201)
+        .json('No estás autorizado para realizar esta acción.');
+    }
   } catch (error) {
     next(error);
   }
